requirements.txt
fastmcp>=0.2.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
googlesearch-python>=1.2.3
pandas>=2.0.0
google-cloud-aiplatform>=1.38.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
gradio>=4.0.0
